refactor(l2): split zkvm-interface into files (#2969)
**Motivation**

A single lib.rs file defining several independent modules is harder to
follow. When new modules are added, this will become worse.

**Description**

This simply moves the modules of lib.rs into separate files. Since they
were already split up, there are no other changes required.
